Copy auto-generated cluster-role, webhook and crd files to helm subrepo
This commit fixes a problem done when replacing Kustomize with Helm in
the project. It didn't copy the files generate during `make manifests`
to the local sub-repo, so when calling `make deploy` the local changes
didn't apply to the cluster.

This commit copy the generated files to the local sub-repo. We add
Kustomize patch during the webhook generation to configure the webhook
service name and add a template variable as service namespace.