Makefile now generates manifests directly in Helm sub-repo (#75)
* fix make undeploy

* delete redundant crd file

* Copy auto-generated cluster-role, webhook and crd files to helm subrepo

This commit fixes a problem done when replacing Kustomize with Helm in
the project. It didn't copy the files generate during `make manifests`
to the local sub-repo, so when calling `make deploy` the local changes
didn't apply to the cluster.

This commit copy the generated files to the local sub-repo. We add
Kustomize patch during the webhook generation to configure the webhook
service name and add a template variable as service namespace.

* Update README.md

Co-authored-by: Ori Shoshan <ori@otterize.com>

* Update src/operator/config/README.md

Co-authored-by: Ori Shoshan <ori@otterize.com>

Co-authored-by: Ori Shoshan <ori@otterize.com>