Use safe RFC1123 for network policy names (#515)
Network policy names are based on the ServiceIdentity name they are created from. The service identity name might contain underscores, which are invalid characters for k8s objects (which must be valid DNS names according to RFC1123). Instead, use a name with the underscores replaced. The implementation of ServiceIdentity is not changed since that might have far-reaching effects in the system.