Default allow metrics-collector netpol for Pods
We want to automatically enable (based on configuration) metrics
collections. This means that is a pod has annotation that marks
Prometheus (and thus also VictoriaMetrics) to scrape metrics on a
specific port - we will create a network policy that allows traffic to
this port.
Since we don't have a deterministic way to identify the
metrics-collector-server - we don't specify any source in the network
policy.