Now handle correctly the service use case
When we notice an change in service \ endpoint resource - we want to
apply the network policy changes in two levels:
1. We want to apply this in the service level - the service itself might
   have the labels indicating for Prometheus scrape. In this case we
will take a look on all the pods that originated from this service - and
if the service is marked for metrics scraping and a network policy need
to be created - we will create it.
2. The other level we need to apply is the pod level - checking its
   annotations for scraping metrics. In case the metrics-collections
configurations is set to "Only If Blocked By Otterize", we want to
create a policy that enables the metrics scraping only if there is
another network policy that blocks this pod. A network policy that might
block the pod could be on the deployment ("the pod"), or on a service
that represents the pod. So when a change in an endpoint is made - we
need to check whether it had an affect on the network policies related
to the pod.

The reason we have a network policy per level (Pod, Service) and not one
for both - is for better management of the ownership. It is easier and
less likely to have errors if we only have 1 owner.