Delete network policy only of the owner instance is the sole owner
We don't want to delete the network policy if another instance is the
owner of the network policy.
For example - if case there is a pod, A, that is the owner of the network
policy, and this instance is being killed - a new pod will raise, B.
When pod B raises - it set the owner of the network policy to itself.
Pod A might finish the termination after pod B already set itself as the
owner of the network policy, so we don't want that pod A termination
will now delete the network policy (there is another living owner).