Auto allow metrics collection based on pod annotations
This is the first layer in managing Prometheus scrape annotations.
When a pod changes, we aim to reduce the entire state of the pod’s
namespace and ensure that the current state of the cluster aligns
with the expected configuration.

We reduce the state of the entire namespace, rather than just the
individual pod, because sometimes we cannot determine everything
needed by looking at just the pod. For example, in the case of a
deployment with a single pod (which has pod annotations for scraping
metrics), when we terminate this pod, there is a race condition
between creating (or keeping) or deleting the pod's
metrics-collection network policy. This depends on when the original
pod terminates.

Instead of attempting to handle many edge cases, we opted for a
stateless approach similar to the one used in
service-effective-policy.
This approach calculates the state and only updates what is necessary.