Allows marking & targeting specific annotation levels for network policies
In Prometheus, we can choose which pod to scrape using scrape
annotations. These annotations can be applied to the pod, a service,
an ingress, and so on. When creating a network policy for scraping
metrics, we aim to cover all possible levels and differentiate
between them.
While we could create a single network policy that handles all
annotations, managing each one separately makes the code more
readable and reduces the number of edge cases. This refactor enables
us to specify and target the annotation level for which we created
the network policy.