Align network policy port decision with Prometheus behavior
If a resource does not specify the `prometheus.io/port`, Prometheus
will attempt to scrape all the ports of the resource.
We have implemented the same behavior for determining the network
policy ports to ensure Prometheus functions correctly.
However, it's important to note that this is not considered a best
practice, and you should explicitly define the `prometheus.io/port`.