Receive metrics scraper server configuration
Since marking a resource in Kubernetes for metrics scraping is a common
and may affect many resources, we aim to make the network policy more restrictive.
To achieve this, we obtain a configuration that identifies the metrics scraping
server. This allows us to create a network policy that only permits traffic between
the metrics scraping server and the deployment.
If we don't receive this configuration, we treat the situation as if automatic
enabling of metrics collection traffic is set to off.