Reconcile on all objects that may affect auto-allow-webhooks
This is why we need to reconcile each of them:
1. CRDs - for conversion webhooks
2. Endpoints - to check whether a pod is blocked by Otterize
3. Services - to check whether a pod is blocked by Otterize
4. Mutating webhooks - for mutating webhooks
5. Network policies - to check whether a pod is blocked by Otterize

A bit more about "to check whether a pod is blocked by Otterize":
In case the configuration is set to "IfBlockedByOtterize", which is the
default, we will create the new "allow-webhook" network policy only if
we see that a workload would be blocked by Otterize network policy.
So we want to be triggered when a network policy is created (in case
this is an Otterize network policy), or when a service is up and
running.