Support DNS traffic when creating metric-traffic network policies
We know that the service that has the Prometheus scraping annotation
might by the DNS service in the cluster (for example, the `coredns`
workload on an EKS cluster).
In this case, adding a network policy to allow metrics-scraping-traffic
would block the DNS ingress traffic, which would block DNS in the
cluster.
We don;t want this, and we can't tell for sure which workload is the DNS
in the cluster, so whenever we create a network policy to allow
Prometheus to scrape metrics, we also allow port 53 (DNS) on the same
network policy